vk: split layout cache from descriptorSet cache
This is just splitting the layout cache into its own class and
files.
